+++ grep-food/Controllers/HomeController.cs
@@ -26,30 +26,13 @@ namespace grep_food.Controllers
 
         public IActionResult Index()
         {
-            var data2 = _dataRepository.Query<RecipeViewModel>().OrderBy(recipe=>recipe.Name).First();
-            //Console.WriteLine($"data: '{data2}' data-siz '{data2.Id}'");
+           
             
             
-            return Recipe(data2.Name);
-
-        }
-        public IActionResult Recipe(string name)
-        {
-            Console.WriteLine($"Retzipe neim: '{name}'");
-            var data = _dataRepository.Query<RecipeViewModel>().Take(10).ToArray();
-            return View(data.Select(x => new RecipeViewModel
-            {
-               
-                Id = x.Id,
-                Name = x.Name,
-                TimeMinutes = x.TimeMinutes,
-                Instructions = x.Instructions,
-                Image = x.Image
-
-            }));
-             
             return View();
+
         }
+     
 
         public IActionResult Privacy()
         {
@@ -60,7 +43,7 @@ namespace grep_food.Controllers
         {
             return View();
         }
-
+       
         [ResponseCache(Duration = 0, Location = ResponseCacheLocation.None, NoStore = true)]
         public IActionResult Error()
         {
