+++ grep-food/Views/Home/Index.cshtml
@@ -63,7 +63,7 @@
                         <a class="nav-link text-uppercase text-expanded" href="Home/Privacy">Products</a>
                     </li>
                     <li class="nav-item px-lg-4">
-                        <a class="nav-link text-uppercase text-expanded" href="store.html">Store</a>
+                        <a class="nav-link text-uppercase text-expanded" href="Recipe/Index?id=36ccce30-7953-4225-bf76-90fe74fd3660">Recipe</a>
                     </li>
                 </ul>
             </div>
