+++ grep-food/Views/Recipe/Index.cshtml
@@ -1,6 +1,10 @@
-﻿@*
-    For more information on enabling MVC for empty projects, visit https://go.microsoft.com/fwlink/?LinkID=397860
-*@
+﻿@{
+    ViewData["Title"] = "Recipe";
+    ViewData["ShowButtons"] = true;
+
+
+}
+@*@model IEnumerable<grep_food.Models.BaseIngredientViewModel>*@
 <html>
 
 
@@ -21,17 +25,17 @@
 
     Hello werdl
     <table>
-    @foreach (var item in Model)
-    {
-        <tr>
-            <td>@item.Name</td>
-        </tr>
-    }
+        @*@foreach (var item in Model)
+        {
+            <tr>
+                <td>@item.Name</td>
+            </tr>
+        }*@
 
-        </table>
+    </table>
 </body>
 
 </html>
 @{
-    
+
 }
